requirements.txt
Pillow==5.1.0
zbar-py==1.0.4
qrcode==6.0
